Handwritten_OCR-main/requirements.txt
absl-py==1.2.0
aiosignal==1.2.0
albumentations==1.2.1
altair==4.2.2
annotated-types==0.5.0
apache-beam==2.37.0
appdirs==1.4.4
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.11.1
biopython==1.79
bleach==3.3.1
blinker==1.6.1
blis==0.7.10
boto3==1.26.114
botocore==1.29.114
cachetools==5.0.0
catalogue==2.0.9
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.6
compress-json==1.0.8
confection==0.1.1
constantly==15.1.0
contextlib2==21.6.0
contourpy==1.1.0
crcmod==1.7
cryptography==41.0.2
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
Cython==0.29.33
d2l==0.17.0
DateTime==5.0
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.6.0/de_core_news_sm-3.6.0-py3-none-any.whl#sha256=cdded59f89eacf348c30a79f9a2485ab21e154ef300cc02c7ebe87dea149c457
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.1.1
distlib==0.3.5
docopt==0.6.2
docutils==0.19
einops==0.2.0
ELM==1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.0.4
fastavro==1.4.10
filelock==3.8.0
Flask==2.0.3
flatbuffers==22.9.24
fonttools==4.28.5
frozenlist==1.3.1
fsspec==2023.4.0
gast==0.4.0
gdown==4.4.0
gensim==3.6.0
getmac==0.8.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-nucleus==0.6.0
google-pasta==0.2.0
GPUtil==1.4.0
greenlet==2.0.1
grpcio==1.43.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.1.0
hdfs==2.7.0
httplib2==0.19.1
huggingface-hub==0.13.4
hyperlink==21.0.0
idna==3.3
ilearnplus==1.4
imageio==2.21.2
imbalanced-learn==0.9.1
imgaug==0.4.0
importlib-metadata==4.11.3
importlib-resources==6.0.1
incremental==22.10.0
install==1.3.4
intel-openmp==2022.2.0
intervaltree==3.1.0
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipython-sql==0.4.1
ipywidgets==7.6.3
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==1.0.1
joblib==1.3.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
langcodes==3.3.0
libclang==14.0.6
lightgbm==3.3.5
llvmlite==0.40.1
lmdb==1.4.1
lxml==4.9.1
Markdown==3.3.6
markdown-it-py==2.2.0
markdownify==0.11.6
MarkupSafe==2.0.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
metric-learn==0.6.2
MiniSom==2.3.0
mistune==0.8.4
mkl==2022.2.0
mock==4.0.3
mpmath==1.3.0
msgpack==1.0.4
murmurhash==1.0.9
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.8.6
nltk==3.8.1
notebook==6.4.0
nucleus-python==0.3.0
numba==0.57.1
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openpyxl==3.0.9
opt-einsum==3.3.0
orjson==3.6.7
outcome==1.2.0
packaging==21.0
pandas==1.5.3
pandocfilters==1.4.3
parsel==1.8.1
parso==0.8.2
pathy==0.10.2
pickleDB==0.9.2
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.2
plot-keras-history==1.1.38
ply==3.11
portalocker==2.7.0
prefetch-generator==1.0.1
preshed==3.0.8
prettytable==0.7.2
prince==0.9.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
Protego==0.2.1
protein-bert==1.0.1
proto-plus==1.20.3
protobuf==3.19.4
psutil==5.9.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
PyDispatcher==2.0.7
pydot==1.4.2
pyfaidx==0.7.2.1
pygad==2.16.3
Pygments==2.15.0
pymongo==3.12.3
Pympler==1.0.1
pynndescent==0.5.10
pyOpenSSL==23.2.0
pyparsing==2.4.7
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
pytorch-transformers==1.2.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pyvi==0.1.1
PyVirtualDisplay==3.0
PyWavelets==1.3.0
pywin32==301
pywinpty==1.1.3
PyYAML==6.0
pyzmq==22.1.0
QDarkStyle==3.1
qtconsole==5.1.1
QtPy==1.9.0
qudida==0.0.4
queuelib==1.6.2
ray==1.13.0
regex==2023.3.23
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==13.3.3
rsa==4.8
s3transfer==0.6.0
sacremoses==0.0.53
sanitize-ml-labels==1.0.50
scikit-image==0.19.3
scikit-learn==1.3.0
scipy==1.10.1
Scrapy==2.9.0
seaborn==0.12.0
selenium==4.7.2
Send2Trash==1.7.1
sentencepiece==0.1.98
service-identity==23.1.0
shapely==2.0.1
sip==6.7.6
six==1.15.0
sklearn==0.0
sklearn-crfsuite==0.3.6
skorch==0.14.0
smart-open==6.3.0
smmap==5.0.0
smote-variants==0.7.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.43
sqlparse==0.4.3
srsly==2.4.7
statistics==1.0.3.5
streamlit==1.21.0
support-developer==1.0.5
sympy==1.12
tabulate==0.8.10
tbb==2021.7.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.10.0
tensorflow-addons==0.20.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.6.0
tensorflow-io==0.24.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
thinc==8.1.12
threadpoolctl==2.2.0
tifffile==2022.8.12
tldextract==3.4.4
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1+cu117
torch-cluster==1.6.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse==0.6.15
torch-spline-conv==1.2.1
torchaudio==2.0.2+cu117
torchdata==0.6.1
torchsummary==1.5.1
torchtext==0.15.2
torchvision==0.15.2+cu117
tornado==6.1
tqdm==4.62.2
traitlets==5.0.5
transformers==4.28.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typeguard==2.13.3
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3
umap-learn==0.5.3
undetected-chromedriver==3.1.7
urllib3==1.26.8
validators==0.20.0
vietocr==0.3.11
virtualenv==20.16.3
vis==0.0.5
w3lib==2.1.1
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
websockets==10.4
Werkzeug==2.0.3
widgetsnbextension==3.5.1
wrapt==1.12.1
wsproto==1.2.0
xgboost==1.6.2
xgboost-ray==0.1.9
zipp==3.7.0
zope.interface==5.5.2
